Refactor transcription module exports
